Primeros Test corriendo.

+++ src/main/java/com/sdc/controller/MovieDAO.java
@@ -5,7 +5,7 @@ import com.sdc.controller.BackEntities.Movie;
 import org.springframework.dao.DataAccessException;
 
 public interface MovieDAO {
-    public Movie getMovie(Long movieId) throws DataAccessException;
+    public Movie getMovie(int movieId) throws DataAccessException;
     public int addMovie(Movie movie);
     public String addActor(ActorB actor);
 }
